Fixes small bugs in gaussian_analytical_coeffs method

+++ csrc/core.cpp
@@ -124,7 +124,7 @@ nfft_fastsum(
 
 torch::Tensor
 gaussian_analytical_coeffs(
-    const float sigma,
+    const double sigma,
     const int64_t N,
     const int64_t dim)
 {
